Make sure the toString methods just call Jackson directly, to avoid double trouble
